+++ saleor/product/models/__init__.py
@@ -1,0 +1,5 @@
+from base_products import *
+from discounts import *
+from products import *
+from variants import *
+from images import *
